oeqa/runtime_test: Disable test_testimage_virgl_gtk
This test keeps failing on the autobuilder and is proving extremely
annoying. It works much better in later releases but for zeus and
earlier, lets just stop running it as it doesn't really tell us
anything useful at this point, nobody has any plans to improve
the distro exclusions or otherwise fix it in the older releases.

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>